Content type filter fixes
