+++ libs/movie/src/lib/movie/form/main/main.component.html
@@ -40,7 +40,7 @@
         </static-select>
       </article>
 
-      <ng-container *ngIf="contentType | async | isType: 'tv'">
+      <ng-container *ngIf="form.contentType.value === 'tv'">
         <article flxLayout="row" fxLayoutAlign="start start" fxLayoutGap="24px">
 
           <mat-form-field appearance="outline" fxFlex>
