+++ libs/movie/src/lib/movie/form/main/main.component.ts
@@ -25,7 +25,6 @@ import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-ti
 export class MovieFormMainComponent implements OnInit, OnDestroy {
   form = this.shell.getForm('movie');
   public movieId = this.route.snapshot.params.movieId;
-  contentType = this.form.contentType.valueChanges.pipe(startWith(this.form.contentType.value))
   valuesCustomGenres$: Observable<string[]>;
   customGenreCtrl = new FormControl();
   public separatorKeysCodes: number[] = [ENTER, COMMA];
