+++ libs/ui/src/lib/form/content-type/content-type.component.html
@@ -1,5 +1,4 @@
-<mat-button-toggle-group test-id="contentTypeToggleGroup" name="contentType" [formControl]="form" fxLayout fxLayoutGap="8px" >
-  <mat-button-toggle value="movie">Movie</mat-button-toggle>
-  <mat-button-toggle value="tv">TV</mat-button-toggle>
+<mat-button-toggle-group name="contentType" [formControl]="form" fxLayout fxLayoutGap="8px" >
+  <mat-button-toggle *ngFor="let type of contentTypes" [value]="type">{{contentType[type]}}</mat-button-toggle>
 </mat-button-toggle-group>
